move diff text formatting into model
